python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "logbull",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    "structlog>=25.4.0",
    "uv>=0.8.22",
]

[tool.uv.sources]
logbull = { path = "../../libraries/python" }
